recovered repository
the system time on the pc I was working on was set to 5 months in the
future. this seemed to cause bugs with the git repository when I synced
it.
this version has the Person class removed but there some code that needs
to be adjusted to get it working right again.